requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
django-ratelimit==4.1.0
psycopg[binary]==3.2.3
redis==5.2.1
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
bcrypt==4.2.1
cryptography==44.0.0
Pillow==11.0.0
python-decouple==3.8
gunicorn==23.0.0
sentry-sdk==2.19.2
drf-spectacular==0.28.0
phonenumbers==8.13.52
requests==2.32.3
dj-database-url==2.3.0
whitenoise==6.8.2
